Add metrics infrastructure
This adds a new metrics infrastructure that's parallel to the existing logging
infrastructure and is available through config.metrics .

Why? Because it makes it much much much easier to thread metrics throughout the
app without adding more tangled component infrastructure and/or detangling the
existing stuff.

This is intentionally a minimal implementation. If we need more, we can break it
out into a larger thing at that point.