Fixes bug 1370484 - Refactored bugzilla query to use REST API and JSON. (#3809)
* Fixes bug 1370484 - Refactored bugzilla query to use REST API and JSON.
This query used to be a very long, hand-written URL, using the old bugzilla query system that returned CSV data. With this, it now uses the REST API instead, manipulating JSON data with the requests library. The query is much simpler to read and has only the few arguments it actually needs. Also, the query is not configurable anymore, because that option was never used and it would have made the code less maintainable.

* Added a clear exception when bugzilla fails.

* Use requests.Response.raise_for_status for HTTP errors.