Fixes bug 1372999 - add document size metrics for elasticsearch crashstorage (#3816)
* Add metrics infrastructure

This adds a new metrics infrastructure that's parallel to the existing logging
infrastructure and is available through config.metrics .

Why? Because it makes it much much much easier to thread metrics throughout the
app without adding more tangled component infrastructure and/or detangling the
existing stuff.

This is intentionally a minimal implementation. If we need more, we can break it
out into a larger thing at that point.

* Fixes bug 1372999 - Add document size metrics for Elasticsearch

This change captures raw_crash and processed_crash sizes for Elasticsearch so we
can see median, 95%, and max over time and better plan Elasticsearch changes.

* Fix nits