Fixes bug 1375551 - install git into circle environment
Ok, so I think what's going on is that our circle configuration downloads an
Ubuntu 14.04 image, installs some packages in it, and then proceeds with the
steps.

On the "checkout:" step, Circle notices that "git" isn't in the container, so
Circle "helpfully" does a git checkout elsehwere, then does a docker cp into the
container, then proceeds with things.

The problem with that is that the container probably has contents from a
previous run. So any files deleted in a PR don't get removed from the container
because docker cp only copies files--it's not syncing files.

The fix here is to install git as a package in the container. Then when it does
the "checkout:" step, it's using git in the container and that will update the
working tree based on the commit correctly and there's no docker cp involved.

Pretty sure that's what's going on. If we bump into this problem again, we can
shake our head in sadness and throw this theory out the window. If that happens,
I suggest we all get religion because I claim the set of possible theories left
will be very small and mostly in the mystical/cosmic domain.