bug 1378836 - Make stackwalker detect likely-wrong crash addresses
On x86-64, when a non-canonical address is referenced
(https://en.wikipedia.org/wiki/X86-64#Canonical_form_addresses) the CPU raises
a General Protection Fault that does not contain the faulting address, so
operating systems put misleading data into the error structures. On Windows
the crash address shows up as `0xFFFFFFFFFFFFFFFF`, on Linux and OS X it shows
up as `0`.

Unfortunately this can make potentially security-sensitive crashes look like
harmless NULL-pointer dereferences, or otherwise confuse developers.

Finding the correct address is hard, but we can at least detect when the
crash address is likely to be wrong and add a note to the output to give
developers a hint.