Fix handling of "no, don't drop the db"
This fixes the logging to use the right method (.warning() vs. .warn()).

This also fixes the socorro script to return the non-zero exit code if there is
one as an exit code. That allows run_setup_postgres.sh to drop out.

Fixes the makefile dockersetup rule to stop if there's an error and updates the
comment for that rule which doesn't look right anymore.