fixes bug 1386781 - fix issues in run_setup_postgres.sh (#3885)
* bug 1386781 - fix issues in run_setup_postgres.sh

This moves the alembic_config to docker_common.env so it's available in both the
processor and webapp containers.

This adds a urlwait to docker/run_setup_postgres.sh to force it to wait for
postgres db to be ready for connections before trying to setup tables and run
migrations.

This moves the database name to a variable that exists already so it's not
hard-coded.

This removes the code in setupdb_app that prevents logging lines from happening
because that makes understanding what happened and debugging it much much
harder.

* Add -e and adjust run_setup_postgres.sh script

This changes the script so that it always drops the db and creates a new one.
Further, it will die if it errors out. It also updates the comments.

* Fix handling of "no, don't drop the db"

This fixes the logging to use the right method (.warning() vs. .warn()).

This also fixes the socorro script to return the non-zero exit code if there is
one as an exit code. That allows run_setup_postgres.sh to drop out.

Fixes the makefile dockersetup rule to stop if there's an error and updates the
comment for that rule which doesn't look right anymore.