fixes bug 1389218 - redo behavior configuration in docker environment (#3914)
* fixes bug 1389218 - redo behavior configuration in docker environment

This redoes how behavior configuration works. Previously, docker-compose would
inject the behavior configuration into the docker container environment. It can
do this because it's being run in a working directory that has the socorro code
in it, so the necessary env files exist *outside* the container it's running.

When we go to run these containers in a server environment or any environment
where you don't have the socorro repo as the working directory, then the
behavior configuration needs to be pulled in from *inside* the container.

This fixes the run_ scripts so that if an indicator flag isn't in the
environment, it pulls the appropriate env file in and executes the component in
the new environment with the behavior configuration.

* Add missing script

* Fix build_env.py to support multiple env files