Redo weekly table creation
This drops the two crontabber jobs and replaces them with a script we can throw
away at some point when postgres is no longer used for crash storage.

The script creates the last 8 weeks of tables plus the next 2 weeks of tables.
That should be a sufficient amount of time that covers crashes we might be
debugging.