Fix stackwalker execution (#3934)
* Fix stackwalker execution

This fixes things so that stackwalker runs successfully

* Clean up BreakpadStackwalkerRule2015 environment variables

This removes the variables we don't need and makes sure the ones we do need are
in the right place with a big comment about them.

* Add scripts/process_and_quit.sh

This runs the processor and tells it to process whatever is in the queue and
then quit. This is useful for automation scripts.