Change all keyword fields to include ignore_above setting
In migrating all the data from the ES 1.x cluster to the ES 5.x cluster, we
observed the following:

1. seems like Winsock_LSP is the only field that has valid values that exceed
   the limit
2. the other fields occasionally have values that exceed the limit, but it's
   junk values from corrupted crash reports

Given that, seems like it's ok to drop values from non-Winsock_LSP fields
because they're junk. Seems like it's ok to drop overly long Winsock_LSP values.
If that turns out to be a bad idea, we should truncate them in the processor
before saving them to Elasticsearch. We can do that in a follow-up fix.