fixes 1385970 signature cli (#3960)
* fixes bug 1385970 - implement signature generation cli

This implements a signature generation cli that lets you use and test signature
generation in isolation from the rest of the Socorro processor.

Usage:

    python -m socorro.signature CRASHID [CRASHID ...]

It also has a verbose mode which lets you see how the signature changes
rule-by-rule which can help when debugging the algorithm.

* bug 1383113 - implement treelib

* Get tests working; remove configman requirement

* Code cleanup

* Squash CSignatureTool classes; code cleanup

* Fix OOMSignature rule

The rule checked to see if it was in the raw crash at all. That's problematic
because it allows for ever shifting raw crash shape. If we ever switched to
using a schema and filling in missing values, that'd break.

* Fix setup for SignatureGenerationRule

This fixes the processed_crash_minimal where the thread data was wrong, so then
the SignatureGenerationRule didn't work for any crashes where the crashing
thread wasn't 0. Oops.

* Remove notes argument from predicate

* Add doc-string

* Implement outputters and CSV output

* Fix watchdog rule

* Add note about classifications

* Implement SOCORRO_API_TOKEN use and add better help text

* Add default keyword to tree_get calls

* Cleanup and docs

* Add signature generation docs to manual

* Fix debug logging and a few other issues

* Cleanup based on PR comments

* fix imports in mozilla_transform_rules.py
* improve treelib per Peter's suggestions
* nix file redirection in cli example
* Ravn DSN -> Sentry DSN
* change behavior of empty paths and edges in tree_get
* remove some unnecessary code
* handle case where processed_crash might not have a signature key
* handle 429 and 500 when retrieving raw and processed crashes
* add tests for SignatureGenerator