fixes bug 1397857 - fix calling format (#3972)
For boto connections, we use RegionalS3ConnectionContext and subclasses. This
family of connection contexts default to OrdinaryCallingFormat rather than
SubdomainCallingFormat. This is important because we have periods in our bucket
names and thus SubdomainCallingFormat doesn't work.

However, when I picked the default for resource.boto.calling_format in the
original commit, I looked at S3ConnectionContext rather than
RegionalS3ConnectionContext and thus used the wrong one and now everything is
hosed in -stage and Sentry is giving me dirty looks late on a Friday just before
the weekend.

This fixes all of that.