Adjust run_webapp.sh text to clarify url
The django runserver tells you to connect at a url formed by taking the host and
port the server is listening to. That's great, however we set it to listen to
0.0.0.0 (all interfaces) and thus the url it says to use is wrong. The code that
generates that message is deep within the bowels of the runserver code, so it's
not really changeable.

This adjusts the message that docker/run_webapp.sh spits out so it's more
prominent and clearer which url to use.