Fix tests running in Circle CI
In Circle, we run the tests using docker-compose, so AWS_HOST and AWS_PORT are
set which means that the symbol upload view uses boto.connect_s3 instead of
boto.s3.connect_to_region.

However, the tests in question mock boto.s3.connect_to_region, so they all break
in an exhilerating fizzle of "service unknown" errors and assertion failures in
the mocking.

This fixes that by setting AWS_HOST to "" when running tests.