fixes bug 1385966 - fix UnicodeEncodeError in AbortSignature
This rule was adding text from who-knows-where to the signature. Since
signatures are used for bucketing, it seemed prudent to convert the string to
something that works well for bucketing. Towards that we now convert the abort
message to a string, drop all unicode characters, drop non-printable characters,
and drop all non-space whitespace characters.