bug 1402037 - improve shutdownhang signatures
We have a number of shutdownhang signatures where the stack is rife with
"uninteresting" bits such that the signature gets truncated before getting to
something interesting or relevant or actionable.

This adds a bunch of things to the irrelevant list such that they don't show up
in the signature.

I did some regression testing and these strings only show up in shutdownhang
crashes. If that changes, we should rethink them on a string-by-string basis.