fixes bug 1403246 - implements paging in fetch_crashids.py
Super Search lets you fetch at most 1000 results in a query. This fixes
fetch_crashids.py so that you can specify a maximum number of results to get or
"all". In both cases, it'll run as many Super Search queries as it needs to
fetch all the results you want.

This further fixes the code to retry on 429s which is handy.

This also also fixes the code to be more resilient of errors. Instead of losing
everything because of an error and having to do it again, you only lose what
hasn't been requested, yet.