bug 1405679 - add submitter as a docker service
This adds the submitter as a docker service with configuration and a run script.
This is different than other services because the configuration conflicts with
configuration for other services.

For now, I'm just slapping this together for the purposes of being able to test
changes to the submitter in the local development environment. If we decide to
use this for other purposes, we can redo it then.