fixes bug 1408054 - fixes bash prompt in containers
When running a command in a container, this will now show the username and
container name in teh bash prompt. For example:

   app@processor

When using docker/as_me.sh which runs the container using a uid and gid of a
user on the host, this will now show "you" as the username. For example:

   you@processor

This will make it much easier to use containers for development.