Remove unused rule predicates and actions
transform_rules had a section for 'useful rule predicates and actions' dating back 5 years. Only one of the methods was referenced outside of testing the methods directly. That reference is implicit, in an alembic migration as the fully qualified path to the method. We no longer use postgres for storing rulesets. I did some archeology to PR 1257 and bug 874505, and can confirm there are rules in the current architecture that implement this functionality without using the DB or referencing these specific methods.