Remove skunk and support classifiers
the skunk and support classifiers powered user emails and experimental features. These days, they only apply to a very small fraction (<0.001% and <1% respectively) of crashes and GA has no record of anyone querying them in supersearch in the last year.

TODO: remove the skunk and support fields from super search fields and the crash report schema

note: to remove these, some tests were modified to rely on other classifiers. the impacted tests were superficial and are likely to be refactored or removed as part of fixing the meta-issue.