fixes bug 1408448 - stomp on signature when parentBuildID != childBuildID
If MozCrashReason is "MOZ_RELEASE_ASSERT(parentBuildID == childBuildID)", then
stomp on the signature with "parentBuildID != childBuildID". This allows all
these crashes to end up in the same bucket where the original signature is
likely to be bogus since there are two different builds involved.

This also adds raw_crash['MozCrashReason'] to the list of things that we use to
generate signatures.