fixes bug 1377254 - clean up CI code
This nixes some files we don't need anymore and switches the test scripts to run
pytest instead of webapp-django/bin/ci.sh.

This removes the ad-hoc linting we had in the webapp code and replaces it with
flake8 that's run in Circle, Travis, and local development environments.

This adds a few more directories to exclude for flake8.