Fix bug 1412281: Update how we configure flatpickr.
Flatpickr now returns a list of selected dates instead of a single date object,
so the onChange callbacks have been updated to unpack the list being returned.
The callbacks themselves were also moved to the initial config object, although
this isn't strictly necessary.

The .eslintrc config was also updated to support the object function shorthand
being used in this patch.