fixes bug 1410284 - prune migrations
This prunes the alembic migrations down to the last few that we've done and
deletes all the other ones. As the tables and code has changed over time, many
of the migrations can't run with the current code base. So keeping them around
isn't helping.

If we ever need to look at migrations for code archaeology, we can look at them
in git history.