fixes bug 1388130 - redo crontabber jobs configuration
This redoes crontabber jobs configuration such that it can take a Python dotted
path to a string with the configuration.

In order to do that, we also had to redo things so that CronTabberApp was a
valid socorro app. This means it now works with the socorro script.

This updates the related docker scripts to use the new way to run it.

In doing all that, I discovered the tests weren't consistent in how they set up
a config manager. This fixes tests, too.