fixes bug 1407655 - implement RawADIMoverCronApp
This implements a third fetch-adi-from-hive cron app. This variation has
a source db and a dest db and copies the data in the raw_adi table over.

This has no unit tests. There is a single integration test that's run
as a shell script by hand and it requires manual verification.