fixes bug 1418037 - add AuroraVersionFixitRule
The AuroraVersionFixitRule checks the crash and stomps on the version if
the crash is from Firefox with specific build ids. The build ids are
for builds that are on the aurora channel for devedition product, but have
a .0b1 beta-style version.

We're fixing the crashes in the processor because the alternative was to
work through the ball of spaghetti that is ftpscraper, the stored procedures,
and related database table data to figure out why it's doing everything WRONG!
The Socorro team therapist suggested we do a processor rule now and rewrite
ftpscraper later and/or switch to buildhub.