Add e2e-tests and system checklist to docs
The e2e-tests README is written in Markdown, so this includes all the things
we need to use Markdown in the docs.

This adds the socorro system test checklist thing we had in my unofficial
socorro-zero repository. We've used variations of it several times now. I'm
sure there are things that are missing, but it's designed to be an evolving
document.