bug 1426225 - capture fetch raw crash errors
The part of the processor that fetches raw crash metadata and dumps and
then stores bits on disk could fail and when it did, it'd log something, but
unless someone was watching the logs at the time, no one would ever know.

This fixes that in the same way we capture other processing errors--it
sends any errors to sentry assuming sentry is configured.