fixes bug 1371428 - nix ' to '' in signature generation
For years, there's been a bit of code to "escape single quotes" in
Java and C signatures. As near as I can tell, it predates the current
era of Socorro requirements and was likely added for some reason that
is no longer relevant.

One consequence is that it makes searching for frames harder, so this
removes that behavior.