Remove flake8 from e2e-tests tox config.
We already run flake8 as part of the automated test suite; this patch enables
that run to also lint the e2e-tests directory, and removes the tox environment
use for running flake8 on the e2e-tests.

However, because we can't have per-directory configuration for flake8, we can't
only ignore line-length limits within the e2e-tests directory anymore, so this
also fixes those errors within e2e-tests.

The upside is that these files will now be linted by our standard automation.