fixes bug 1424659 - nix consecutive spaces in signatures
Prior to this, it was possible (though rare) to have two spaces in a row
in signatures. That's tough because the webui shows that as a single
space and if you search for the single-space version, you won't get any
results.

This fixes that by establishing that signatures can't have more than 1
consecutive whitespace character. Further, this makes sure that all
non-space whitespace characters are converted to space and this
happens all before we truncate the signature.

I also changed SigTrunc to SigTruncate because "trunc" is an unnecessary
abbreviation.