fixes bug 1429863 - add ./scripts/populate_data.sh
This adds a script that will pull down NUM_CRASHES crashes per day for the
last NUM_DAYS days, sync all the crash data to the local S3, and then
send the crash ids to the local rabbitmq queue for processing.

The use case here is to establish a one-liner that populates a local
development environment with the data required for the product home page
and other reports to work.