Fixes Bug 1434495: Update stackwalker to process ModuleSignatureInfo field
As of bug 1436845, Firefox now sends an annotation named "ModuleSignatureInfo"
with crash reports. This field contains json that describes the owners of the
certificates used to sign various binaries in the crash report.

The format is as follows:

{
  "Cert owner": ["list", "of", "dlls", "with", "the", "same", "cert", "owner"],
  "Another One": ["More", "stuff"]
}

This patch modifies the stackwalker to read that annotation and then add the
cert info to each applicable entry in the report's module data.