bug 1422923 - removes symbols upload ui and code
This removes all of the code from the symbols app. It includes a db migration
to drop the symbols_symbolsupload table. It keeps the minimal bits for a
Django app so that the migration runs. We'll keep that for a while and then
remove that, too.

This also creates a redirect from /symbols/* to https://symbols.mozilla.org/
in case people try to go here for symbol uploads.