fixes bug 1436537 - fixes sha to tag with in deploy_bug.py
Previously we removed the merge commits from the list of commits that got
displayed and then figured out the commit to tag by looking at the last commit
in that list. However, that's wrong--we really want the last commit and not
the last non-merge commit especially since the last commit is likely to be
a merge commit.

This fixes that. It also cleans up the code a bit.