fix bug 1435551 - stop saving to processed_crashes_* tables
We use the PostgreSQLCrashStorage as one of the destinations in the
processor. That crash storage class saves raw crash data to the
raw_crashes_* tables and processed crash data to a series of reports
tables and also the processed_crashes_* tables.

However, we no longer look at the data in the processed_crashes_* tables.
As such, it's a waste of work, bandwidth, and db use.

This nixes that extraneous work.

This also merges PostgreSQLBasicCrashStorage with PostgreSQLCrashStorage
since it doesn't help us to have two separate classes and all the tests
and everything.