bug 1440745 - change age lockout to 2 hours
The default lockout is 12 hours. That means if we kill a crontabber node
that's in the middle of doing something, then that job is locked for 12
hours before it can be done.

With the new infrastructure, it's much more likely we're going to kill a
crontabber node in the middle of doing something. We really don't want
anything to not run for 12 hours.

This changes that to 2 hours. That should be plenty long enough for any of
the jobs to run but a much more reasonable amount of time to wait for a
lock to time out.