bug 1443981 - convert some boto tests to use moto
This creates a BotoHelper and pytest fixture. Then it converts the
test_crash_data tests to use that and moto instead of mock.

This also changes the other tests so that they don't subclass unittest
TestCase which prevents us from using pytest fixtures.

There's still work to do, but this moves us forward.