Get rid of super search fields mocks
Super search fields used to be stored in Elasticsearch, so mocking them
was helpful. Now, they're hard-coded in a Python file, so mocking them
isn't helpful anymore and prevents us from writing tests that don't mock
all things Elasticsearch.

This removes the super search fields mocking and fixes tests that break
because of it.