Add two new histograms for elasticsearch
This adds two new histograms, which track the count and timing of
successful and failed saves of crashes to elasticsearch.

The existing histogram for
processor.es.ESCrashStorageRedactedJsonDump.save_raw_and_processed
encompasses a chunk of python code and potentially multiple calls to
ES. This new metrics will get us just the time spent talking to ES and
let us distinguish between successful and failed calls.

This should help us understand the performance of the new
infrastructure.