fix bug 1375434: redo first signature maintenance
This creates a new cron job that looks at data in Elasticsearch and updates
the first_build and first_report data in the signatures table accordingly.

This job is set up to run hourly and look at a 90 minute window. That'll
help it not miss things. It also keeps the Elasticsearch queries small enough
that it can work through them quickly.

This job is set up to be resilient. If it dies in the middle--no big deal.
If there are two running at the same time--no big deal. We get enough crash
data that if we miss some things--no big deal.