fix bug 1460125: remove postgres crashstorage in processor
This takes the postgres crashstorage out of the processor crash destinations.
It also removes the ReportsCleanCronApp since we no longer need that anymore.

This moves some code around to fill the void that removing
PostgreSQLCrashStorage left behind and tidies some things up.

This does not remove related tables and stored procedures which are no
longer needed. Those will get removed in a future PR.