fix bug 1460035: implement python 3 test scaffolding
This adds a couple of make rules and shell scripts to give us scaffolding
with which we can iterate over fixing code to make it work in both
Python 2 and Python 3.

In order to get the tests to run at all, I had to fix a bunch of Python
3 issues. I also hoisted six to first-class dependency status.

This doesn't connect that with CI because linting fails epically. We'll do
that later.