Add ExpiringCache for in-memory cache with a ttl
I looked for libraries and didn't like anything I found and then decided
it's easier to roll our own where this is a very limited scope.

Having a cache where keys have ttl allows us to cache values in the
processor such that they update themselves over the course of the
processor's lifetime without us having to rely on deploys to update the
processor's view of the world.