Fix building permissions and groups in local dev environment
At some point we upgraded Django to a version that didn't work with the way
we were setting up permissions in the local development environment. This
redoes that code and cleans it up a bit. This also adds code to create the
Hackers and Hackers Plus groups which we need.