fix bug 1464107: add script to run migrations
This codifies all the migration stuff in a single script which we'll run
as a post-deploy script.

This adds curl and gawk to the Ubuntu setup. We need curl for installing
sentry-cli and we need gawk because it's the version of awk that sentry-cli
uses.

This also adds the sentry-cli to the crontabber container. This enables
us to send errors in bash scripts to Sentry.