bug 1462749: remove api token management ui
API tokens are Django models and we switched to using the Django admin to manage
admin-centric actions for API tokens.

This does not remove the views for user-centric actions for API tokens. Users
can continue to create, edit, and delete their API tokens using the existing
views and forms--we don't use the Django admin for that.