Remove auditing from graphics_devices views
The graphics_device table is going to get moved to Django-land soon and
we'll be able to administrate it with the Django admin. Also, we're
migrating off of pinax's eventlog to the regular Django admin log.
On top of that, I've been adminning Socorro for a year and never
used this view.

Between those three things, seems like a better idea to just drop the
auditing code for this view for now.