fix bug 1467308: Update socorro.lib folder for Python 3
We were dealing with a pytest that captured an expected error. When we
capture the error, there is an explicit assertion of what the error is
doing and its type. We then cast it to a string to make an assertion
about the value of the error. We then found out the behavior of
representing the class name of the error as a string changed between
python 2 and 3. Therefore to solve this we explicitly by capturing just
the value of the tested error and modified our expected error just to
include the error value expected so we could pass comparison test.
Because we already asserted the type of error, and there was a change of
behavior around representing the error type as a string. We resolved the
difference in python 2 to 3 by removing the redundant verification of
the error type.