Move siglist code into signatures/ directory
This separates the siglist manipulation code from the siglist files.

The siglist files need to be in a directory by themselves with a README to
make it easier for people to do drive-by changes to them. Given that, it
makes sense to move the code up a directory to the signatures directory.