bug 1469718: all the changes for sub dir socorro.external.boto
In testing the raw dump data there was an error in the test value
returning as a binary string. This value was failing in comparison to
the expected unicode value for python 3. Placing print statements in
class level functions didn’t tell us much.So we traced the function from
top to bottom and bottom to top. This pushed over the deep end as we
found ourselves in the boto service source code. We couldn’t find the
exact time this conversion of the test value occurred. To solve this
issue we made the assumption data being used in raw dump is binary. So
we change the test value to be a binary string and also changed the
expected comparison to be a binary string as well. The implementation of
the function iteritems() changed from python 2 to 3. Similar to the
range function the change in implementation went from generating a list
to using an iterator. Lastly objects of type ‘dict_values’ is not JSON
serializable. To solve this we wrap the call to values with a call to
list.