fix bug 1475820: make begin_panic and begin_panic_fmt sentinels
When rust code panics, it goes through a panic! macro and the result of
that is there are a bunch of "oh, I'm panicking!" frames until we get to
the context of the actual panic.

We don't want the "oh, I'm panicking!" frames in signatures, so best to
start signature generation after those. This adds sentinels to fix that.