Remove ProductVersionsCronApp
This is the last matview crontabber app. It's sole purpose is to run
the update_product_versions stored procedure. However, ftpscraper and
the Releases API both run update_product_versions after doing any
changes to products and releases, so having a separate crontabber app
that does the same thing is superfluous now.

This removes that and the matview crontabber app infrastructure.