bug 1477782: implement auditgroups Django command
This implements an auditgroups Django command which does the work of auditing
all the users in the Hackers group and removing anyone that doesn't meet
certain criteria.

This doesn't implement anything to run that audit on a periodic basis. I'm
going to wait until we've sorted out the crontabber situation until we do
that.