fix bug 1478741: add ftpscraper wrapper to make it FASTER
This adds an ftpscraper wrapper to be used in the local development
environment that takes a lot of pain out of slicking-and-rebuilding
a local development environment.

It keeps a lot of ftpscraper and replays it if the log is less than 7 days
old. "make clean" will wipe the log. That covers most of my use cases.