fix bug 1479097: unify sentry captureException uses
This unifies the two Sentry captureException uses into a single
capture_error implementation which sends the error to sentry if possible
or logs an error.

This also fixes a lot of the code and tests that used that so they're
all passing in an exc_info() and not just exceptions. I'm not sure
how some of that code worked--I suspect it didn't.